refactor: productCard의 thumbnail 컴포넌트 분리
